Corrected query builder and tests

+++ src/main/java/com/mealplannerv2/entity/Ingredient.java
@@ -4,7 +4,7 @@ import org.springframework.data.mongodb.core.mapping.Field;
 
 public record Ingredient(
         @Field("name") String name,
-        @Field("amount") float amount,
+        @Field("amount") double amount,
         @Field("unit") String unit
 ) {
 }
